unix: add disk ioctl constants for FreeBSD
These ioctls make it possible to obtain information on the geometry of
disk devices, such as their sector size and total size in bytes.

I was able to generate these constants on each of the architectures in
question, except for GOARCH=arm, as it is not possible to run this port
in an emulator. I therefore copied the values for GOARCH=386.